+++ inst/tinytest/test_vi_shap.R
@@ -25,6 +25,6 @@ vis3 <- vi(fit1, method = "shap", pred_wrapper = pfun, nsim = 10, train = trn1)
 
 # Expectations
 for (v in list(vis1, vis2, vis3)) {
-  expect_inherits(v, c("vi", "tbl_df", "tbl", "data.frame"))
+  expect_inherits(v, c("vi", "data.frame"))
   expect_true(all(c("Variable", "Importance") %in% names(v)))
 }